refactor(ui): move components/form_navigation; finalize components hub shim (S4i) (#102)
