+++ crates/dbflux_ui/src/ui/components/mod.rs
@@ -1,40 +1,15 @@
-pub mod context_menu;
-pub use dbflux_components::components::data_table;
-pub use dbflux_components::components::document_tree;
-pub mod dropdown;
-pub use dbflux_components::components::filter_bar;
-pub mod form_navigation;
-pub use dbflux_components::components::form_renderer;
-pub use dbflux_components::components::json_editor_view;
+pub use dbflux_components::components::{
+    data_table, document_tree, filter_bar, form_navigation, form_renderer, json_editor_view,
+    multi_select, tree_nav, value_source_selector,
+};
+
+// Stays local in dbflux_ui (adapter over dbflux_components::composites):
 pub mod modal_frame;
-pub use dbflux_components::components::multi_select;
-pub mod toast;
-pub use dbflux_components::components::tree_nav;
+
+// Existing shims to other dbflux_components modules (unchanged):
+pub mod context_menu;
+pub mod dropdown;
 pub mod typography;
-pub use dbflux_components::components::value_source_selector;
 
-#[cfg(test)]
-mod tests {
-    use std::fs;
-
-    const COMPONENTS_DIR: &str = concat!(env!("CARGO_MANIFEST_DIR"), "/src/ui/components");
-
-    #[test]
-    fn legacy_surface_modules_are_not_exported() {
-        let source = fs::read_to_string(format!("{COMPONENTS_DIR}/mod.rs"))
-            .unwrap_or_else(|error| panic!("failed to read components/mod.rs: {error}"));
-        let production_source = source
-            .split("#[cfg(test)]")
-            .next()
-            .expect("components/mod.rs should contain production code before tests");
-
-        assert!(!production_source.contains("pub mod surfaces;"));
-        assert!(!production_source.contains("pub mod surfaces_style;"));
-    }
-
-    #[test]
-    fn legacy_surface_source_files_are_removed() {
-        assert!(!std::path::Path::new(&format!("{COMPONENTS_DIR}/surfaces.rs")).exists());
-        assert!(!std::path::Path::new(&format!("{COMPONENTS_DIR}/surfaces_style.rs")).exists());
-    }
-}
+// Deferred:
+pub mod toast;
